+++ jacop-core/src/main/java/org/jacop/constraints/Disjoint.java
@@ -418,7 +418,7 @@ public class Disjoint extends Diff {
           }
           commonArea += partialCommonArea;
         }
-        if (commonArea + r.minArea() > (r_max[0] - r_min[0]) * (r_max[1] - r_min[1])) {
+        if (commonArea + r.minArea() > (long) (r_max[0] - r_min[0]) * (r_max[1] - r_min[1])) {
           throw Store.failException;
         }
       }
